更改JS日期显示BUG

+++ pf_business/src/main/java/com/nnte/pf_business/component/merchant/PlateformMerchanComponent.java
@@ -141,7 +141,7 @@ public class PlateformMerchanComponent extends BaseBusiComponent {
         }
         return ret;
     }
-    public PlateformMerchant getMerchantById(Long id){
+    public PlateformMerchant getMerchantById(Integer id){
         return plateformMerchantService.findModelByKey(id);
     }
